Timesheet api complete and do some changes in project & user so time sheet changes will be there

+++ database/migrations/2025_03_13_120400_create_timesheets_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('project_id')->constrained()->onDelete('cascade');
             $table->string('task_name');
             $table->date('date');
-            $table->integer('hours');
+            $table->decimal('hours', 5, 2);
             $table->timestamps();
         });
     }
